+++ src/arkscript/REPL/Repl.cpp
@@ -1,4 +1,6 @@
 #include <fstream>
+#include <iostream>
+#include <filesystem>
 #include <fmt/core.h>
 
 #include <CLI/REPL/Repl.hpp>
@@ -22,7 +24,6 @@ namespace Ark
         fmt::print("ArkScript REPL -- Version {} [LICENSE: Mozilla Public License 2.0]\nType \"quit\" to quit. Try \"help\" for more informations\n", ARK_FULL_VERSION);
         cuiSetup();
 
-        std::string code;
         while (m_running)
         {
             auto maybe_block = getCodeBlock();
@@ -31,7 +32,7 @@ namespace Ark
             m_old_ip = vm.m_execution_contexts[0]->ip;
             if (maybe_block.has_value() && !maybe_block.value().empty())
             {
-                std::string new_code = code + maybe_block.value();
+                std::string new_code = m_code + maybe_block.value();
                 if (state.doString(new_code))
                 {
                     // for only one vm init
@@ -46,7 +47,7 @@ namespace Ark
                     if (vm.safeRun(*vm.m_execution_contexts[0]) == 0)
                     {
                         // save good code
-                        code = new_code;
+                        m_code = new_code;
                         // place ip to end of bytecode instruction (HALT)
                         vm.m_execution_contexts[0]->ip -= 4;
                     }
@@ -74,15 +75,15 @@ namespace Ark
 
         m_repl.set_word_break_characters(" \t.,-%!;:=*~^'\"/?<>|[](){}");
         m_repl.set_completion_count_cutoff(128);
-        m_repl.set_double_tab_completion(false);
+        m_repl.set_double_tab_completion(true);
         m_repl.set_complete_on_empty(true);
         m_repl.set_beep_on_ambiguous_completion(false);
         m_repl.set_no_color(false);
     }
 
-    std::optional<std::string> Repl::getLine()
+    std::optional<std::string> Repl::getLine(bool continuation)
     {
-        std::string prompt = fmt::format("main:{:0>3}> ", m_line_count);
+        std::string prompt = fmt::format("main:{:0>3}{} ", m_line_count, continuation ? ":" : ">");
 
         const char* buf { nullptr };
         do
@@ -109,6 +110,7 @@ namespace Ark
             std::cout << "  help -- display this message\n";
             std::cout << "  quit -- quit the REPL\n";
             std::cout << "  save -- save the history to disk\n";
+            std::cout << "  code -- print saved code\n";
 
             return std::nullopt;
         }
@@ -118,6 +120,13 @@ namespace Ark
             m_repl.history_save(history_file);
 
             fmt::print("Saved {} lines of history to arkscript_repl_history.ark\n", m_line_count);
+
+            return std::nullopt;
+        }
+        else if (line == "code")
+        {
+            std::cout << "\n"
+                      << m_code << "\n";
 
             return std::nullopt;
         }
@@ -133,20 +142,23 @@ namespace Ark
 
         while (m_running)
         {
-            auto maybe_line = getLine();
-            if (!maybe_line.has_value())
+            bool unfinished_block = open_parentheses != 0 || open_braces != 0;
+
+            auto maybe_line = getLine(unfinished_block);
+            if (!maybe_line.has_value() && !unfinished_block)
                 return std::nullopt;
-            else if (!maybe_line.value().empty())
+
+            if (maybe_line.has_value() && !maybe_line.value().empty())
             {
                 code_block += maybe_line.value() + "\n";
                 open_parentheses += countOpenEnclosures(maybe_line.value(), '(', ')');
                 open_braces += countOpenEnclosures(maybe_line.value(), '{', '}');
-            }
 
-            // lines number incrementation
-            ++m_line_count;
-            if (open_parentheses == 0 && open_braces == 0)
-                break;
+                // lines number incrementation
+                ++m_line_count;
+                if (open_parentheses == 0 && open_braces == 0)
+                    break;
+            }
         }
 
         return code_block;